Task P1C3.1

+++ P1/P1C3.1/Program.cs
@@ -2,7 +2,7 @@
 int denominateur = 4;
 
 // TODO : modifier l'instruction ci-dessous pour que la réponse prenne une valeur décimale
-int reponse = numerateur / denominateur;
+double reponse = numerateur / (double) denominateur;
 
 // Afficher le résultat 
 Console.WriteLine(numerateur + " / " + denominateur + " = " + reponse);
